👍

+++ mycontacts/src/app/controllers/ContactController.js
@@ -1,16 +1,18 @@
 const ContactsRepository = require('../repositories/ContactsRepository')
-
+const isValidUUID = require('../utils/isValidUUID​')
 class ContactController {
     async index (request, response){
         const {orderBy} = request.query;
         const contacts = await ContactsRepository.findAll(orderBy);
-
-
         response.json(contacts)
     }
 
     async show(request, response){
         const { id } = request.params;
+
+        if (!isValidUUID(id)){
+            return response.status(400).json({error : 'Invalid contact id'});
+        }
 
         const contact = await ContactsRepository.findById(id)
 
@@ -22,18 +24,27 @@ class ContactController {
     }
 
     async store(request, response){
-        const {name, email, phone, category_id} = request.body
+        const {name, email, phone, category_id} = request.body;
+
         if(!name){
             return response.status(400).json({error: "Name is required"})
         }
+        if(email){
+            const contactExists = await ContactsRepository.findByEmail(email);
+            if(contactExists){
+                return response.status(400).json({error: "This e-mail is alredy in use"})
+            }
+        }
 
-        const contactExists = await ContactsRepository.findByEmail(email);
-        if(contactExists){
-            return response.status(400).json({error: "This e-mail is alredy in use"})
+        if( category_id && !isValidUUID(category_id)){
+            return response.status(400).json({error : 'Invalid category'});
         }
 
         const contact = await ContactsRepository.create({
-            name, email, phone, category_id
+            name,
+            email: email || null,
+            phone,
+            category_id: category_id || null,
         });
         return response.status(201).json(contact)
     }
@@ -43,22 +54,35 @@ class ContactController {
 
         const {name, email, phone, category_id} = request.body
 
-        const contactExists = ContactsRepository.findById(id);
-
-        if(!contactExists){
-            return response.status(404).json({error:'Contact not found'})
+        if (!isValidUUID(id)){
+            return response.status(400).json({error : 'Invalid contact id'});
         }
+
+        if( category_id && !isValidUUID(category_id)){
+            return response.status(400).json({error : 'Invalid category'});
+        }
+
         if(!name){
             return response.status(400).json({error: "Name is required"})
         }
-        const contactByEmail = await ContactsRepository.findByEmail(email);
 
-        if(contactByEmail && contactByEmail.id != id ){
-            return response.status(400).json({error: "This e-mail is alredy in use"})
+        const contactExists = ContactsRepository.findById(id);
+        if(!contactExists){
+            return response.status(404).json({error:'Contact not found'})
+        }
+        if(email){
+            const contactByEmail = await ContactsRepository.findByEmail(email);
+            if(contactByEmail && contactByEmail.id != id ){
+                return response.status(400).json({error: "This e-mail is alredy in use"})
+            }
+
         }
 
         const contact = await ContactsRepository.update(id, {
-            name, email, phone, category_id
+            name,
+            email: email || null,
+            phone,
+            category_id: category_id || null,
         })
 
         response.json(contact)
@@ -67,6 +91,10 @@ class ContactController {
 
     async delete(request, response){
         const { id } = request.params;
+
+        if (!isValidUUID(id)){
+            return response.status(400).json({error : 'Invalid contact id'});
+        }
 
         await ContactsRepository.delete(id);
